Updated grid array & Bug Fix
Fixed a bug with the flipping logic. Additionally, changed the data structure of the array holding the game grid. The y index now holds an object containing it's attributes as well as a reference to the respective DOM element. This breaks the structuredClone() function I was using before to store default game settings, so I resorted to manually resetting the important settings to their defaults in the gameOver() function.